# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Teleassistance API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/teleoperators": {
            "get": {
                "tags": [
                    "Teleoperators"
                ],
                "summary": "Muestra todos los teleoperadores paginados",
                "operationId": "a819fa4993ecdc315b25c13bd80c4bd2",
                "responses": {
                    "200": {
                        "description": "Lista de teleoperadores",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TeleoperatorResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Teleoperators"
                ],
                "summary": "Crea un nuevo teleoperador",
                "operationId": "cb23446058e5c14e457364819aa32e59",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreTeleoperatorRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Teleoperador creado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "teleoperator": {
                                            "$ref": "#/components/schemas/TeleoperatorResource"
                                        },
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/teleoperators/{id}": {
            "get": {
                "tags": [
                    "Teleoperators"
                ],
                "summary": "Muestra un teleoperador",
                "operationId": "354c0202dd2bf451255fb5d0d7524d0a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del teleoperador",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Teleoperador recuperado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TeleoperatorResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Teleoperators"
                ],
                "summary": "Actualiza un teleoperador",
                "operationId": "31b2619ca623a4727cd3c421ef0242b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del teleoperador",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTeleoperatorRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Teleoperador actualizado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TeleoperatorResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Teleoperators"
                ],
                "summary": "Elimina un teleoperador",
                "operationId": "a40f12020e5d471b4d3d0d6ce07e68ce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del teleoperador",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Teleoperador eliminado con éxito"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreTeleoperatorRequest": {
                "description": "Validación para la creación de teleoperadores",
                "required": [
                    "name",
                    "email",
                    "prefix",
                    "phone_number",
                    "zone_id",
                    "hiring_date",
                    "code",
                    "password",
                    "languages"
                ],
                "properties": {
                    "name": {
                        "description": "Nombre del teleoperador",
                        "type": "string",
                        "maxLength": 255,
                        "example": "María González"
                    },
                    "email": {
                        "description": "Correo electrónico del teleoperador (único)",
                        "type": "string",
                        "format": "email",
                        "example": "maria.gonzalez@example.com"
                    },
                    "prefix": {
                        "description": "Prefijo telefónico",
                        "type": "string",
                        "example": "+34"
                    },
                    "phone_number": {
                        "description": "Número de teléfono del teleoperador (único)",
                        "type": "string",
                        "maxLength": 20,
                        "example": "678901234"
                    },
                    "zone_id": {
                        "description": "ID de la zona a la que pertenece el teleoperador",
                        "type": "integer",
                        "example": 2
                    },
                    "hiring_date": {
                        "description": "Fecha de contratación del teleoperador",
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-15"
                    },
                    "code": {
                        "description": "Código único del teleoperador",
                        "type": "integer",
                        "example": 12345
                    },
                    "password": {
                        "description": "Contraseña del teleoperador",
                        "type": "string",
                        "minLength": 6,
                        "example": "secreta123"
                    },
                    "languages": {
                        "description": "Lista de IDs de idiomas que habla el teleoperador",
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                },
                "type": "object"
            },
...
            "UpdateTeleoperatorRequest": {
                "description": "Validación para la actualización de un teleoperador",
                "required": [
                    "name",
                    "email",
                    "prefix",
                    "phone_number",
                    "zone_id",
                    "hiring_date",
                    "code",
                    "languages"
                ],
                "properties": {
                    "name": {
                        "description": "Nombre del teleoperador",
                        "type": "string",
                        "example": "Carlos Pérez"
                    },
                    "email": {
                        "description": "Correo electrónico",
                        "type": "string",
                        "format": "email",
                        "example": "carlos.perez@example.com"
                    },
                    "prefix": {
                        "description": "Prefijo telefónico",
                        "type": "string",
                        "example": "+34"
                    },
                    "phone_number": {
                        "description": "Número de teléfono",
                        "type": "string",
                        "example": "600123456"
                    },
                    "zone_id": {
                        "description": "ID de la zona asignada",
                        "type": "integer",
                        "example": 2
                    },
                    "hiring_date": {
                        "description": "Fecha de contratación",
                        "type": "string",
                        "format": "date",
                        "example": "2023-07-15"
                    },
                    "code": {
                        "description": "Código único del teleoperador",
                        "type": "integer",
                        "example": 1058
                    },
                    "password": {
                        "description": "Contraseña del teleoperador",
                        "type": "string",
                        "example": "securepassword123",
                        "nullable": true
                    },
                    "languages": {
                        "type": "array",
                        "items": {
                            "description": "ID del idioma",
                            "type": "integer",
                            "example": 1
                        }
                    }
                },
                "type": "object"
            },
...
            "TeleoperatorResource": {
                "description": "Esquema del recurso Teleoperador",
                "properties": {
                    "id": {
                        "description": "Identificador del teleoperador",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nombre del teleoperador",
                        "type": "string"
                    },
                    "email": {
                        "description": "Correo electrónico del teleoperador",
                        "type": "string",
                        "format": "email"
                    },
                    "prefix": {
                        "description": "Prefijo telefónico",
                        "type": "string"
                    },
                    "phone_number": {
                        "description": "Número de teléfono del teleoperador",
                        "type": "string"
                    },
                    "zone": {
                        "$ref": "#/components/schemas/ZoneResource"
                    },
                    "languages": {
                        "description": "Idiomas hablados por el teleoperador",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "hiring_date": {
                        "description": "Fecha de contratación",
                        "type": "string",
                        "format": "date"
                    },
                    "code": {
                        "description": "Código de identificación del teleoperador",
                        "type": "string"
                    },
                    "firing_date": {
                        "description": "Fecha de despido (puede ser nula si sigue activo)",
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ZoneResource": {
                "description": "Esquema del recurso Zona",
                "properties": {
                    "id": {
                        "description": "Identificador de la zona",
                        "type": "integer"
                    },
                    "city": {
                        "description": "Nombre de la ciudad",
                        "type": "string"
                    },
                    "zone": {
                        "description": "Nombre de la zona dentro de la ciudad",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}